Feat: ModelAttribute, DTO로 수정

+++ src/main/java/com/seniors/domain/post/service/PostService.java
@@ -4,8 +4,10 @@ import com.seniors.common.dto.CustomPage;
 import com.seniors.common.exception.type.BadRequestException;
 import com.seniors.common.exception.type.NotFoundException;
 import com.seniors.config.S3Uploader;
+import com.seniors.domain.post.dto.PostDto;
 import com.seniors.domain.post.dto.PostDto.GetPostRes;
 import com.seniors.domain.post.dto.PostDto.ModifyPostReq;
+import com.seniors.domain.post.dto.PostDto.PostCreateDto;
 import com.seniors.domain.post.dto.PostDto.SavePostReq;
 import com.seniors.domain.post.entity.Post;
 import com.seniors.domain.post.entity.PostMedia;
@@ -23,9 +25,13 @@ import org.springframework.data.domain.Sort;
 import org.springframework.data.domain.Sort.Direction;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.validation.BindingResult;
+import org.springframework.validation.FieldError;
+import org.springframework.validation.ObjectError;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.List;
 
 @Slf4j
@@ -40,17 +46,25 @@ public class PostService {
 	private final S3Uploader s3Uploader;
 
 	@Transactional
-	public void addPost(String title, String content, List<MultipartFile> files, Long userId) throws IOException {
-		if (title == null || title.isEmpty() || content == null || content.isEmpty()) {
-			throw new BadRequestException("Title or Content is required");
+	public void addPost(PostCreateDto postCreateDto, BindingResult bindingResult, Long userId) throws IOException {
+		if (bindingResult.hasErrors()) {
+			List<ObjectError> errors = bindingResult.getAllErrors();
+			List<String> errorMessages = new ArrayList<>();
+
+			for (ObjectError error : errors) {
+				FieldError fieldError = (FieldError) error;
+				String message = fieldError.getDefaultMessage();
+				errorMessages.add(message);
+			}
+			throw new BadRequestException(String.join(", ", errorMessages));
 		}
 
 		Users users = usersRepository.findById(userId).orElseThrow(
 				() -> new NotFoundException("유효하지 않은 회원입니다.")
 		);
-		Post post = postRepository.save(Post.of(title, content, users));
-		if (files != null && !files.isEmpty()) {
-			for (MultipartFile file : files) {
+		Post post = postRepository.save(Post.of(postCreateDto.getTitle(), postCreateDto.getContent(), users));
+		if (postCreateDto.getFiles() != null && !postCreateDto.getFiles().isEmpty()) {
+			for (MultipartFile file : postCreateDto.getFiles()) {
 				String uploadImagePath = s3Uploader.upload(file, "posts/media/" + post.getId().toString());
 				postMediaRepository.save(PostMedia.of(uploadImagePath, post));
 			}
@@ -71,17 +85,29 @@ public class PostService {
 	}
 
 	@Transactional
-	public void modifyPost(String title, String content, List<MultipartFile> files, Long postId, Long userId) throws IOException {
+	public void modifyPost(PostCreateDto postCreateDto, BindingResult bindingResult, Long postId, Long userId) throws IOException {
+
+		if (bindingResult.hasErrors()) {
+			List<ObjectError> errors = bindingResult.getAllErrors();
+			List<String> errorMessages = new ArrayList<>();
+
+			for (ObjectError error : errors) {
+				FieldError fieldError = (FieldError) error;
+				String message = fieldError.getDefaultMessage();
+				errorMessages.add(message);
+			}
+			throw new BadRequestException(String.join(", ", errorMessages));
+		}
 		Post post = postRepository.findById(postId).orElseThrow(() -> new NotFoundException("유효하지 않은 게시글입니다."));
-		postRepository.modifyPost(title, content, postId, userId);
+		postRepository.modifyPost(postCreateDto.getTitle(), postCreateDto.getContent(), postId, userId);
 
 		// 기존 미디어 파일 삭제
 		s3Uploader.deleteS3Object("posts/media/" + post.getId().toString());
 
 		postMediaRepository.deleteByPostId(postId);
 
-		if (files != null && !files.isEmpty()) {
-			for (MultipartFile file : files) {
+		if (postCreateDto.getFiles() != null && !postCreateDto.getFiles().isEmpty()) {
+			for (MultipartFile file : postCreateDto.getFiles()) {
 				String uploadImagePath = s3Uploader.upload(file, "posts/media/" + post.getId().toString());
 				postMediaRepository.save(PostMedia.of(uploadImagePath, post));
 			}
